func-data-analysis/requirements.txt
azure-functions
python-dotenv==1.0.1
openai==1.13.3